add as a method + cargo fmt
