+++ sample-config/default-config.toml
@@ -58,6 +58,7 @@ prune_history = 0
 
 # term_command = "alacritty -e {}" # command for applications run in terminal (default uses "$TERMINAL -e")
 # set_gpu_variable = "PRIME" # variable to set to offload if specified by application (default doesn't offload)
+
 # specify name overrides (id is the name of the desktop file)
 [name_overrides]
 # id = "name\rextra"
