+++ src/app_entry/desktop_entry.rs
@@ -1,7 +1,9 @@
 use freedesktop_desktop_entry::{default_paths, get_languages_from_env, Group, Iter};
-use log::{info, error, warn};
+use log::{error, info, warn};
+use shlex::Shlex;
 use which::which;
 
+use std::process::{id, Command};
 use std::{collections::HashMap, env::var, path::PathBuf};
 
 pub struct DesktopEntry {
@@ -127,6 +129,82 @@ impl DesktopEntry {
             }
         }
         out.into_values().collect()
+    }
+    pub fn run(
+        &self,
+        term_command: Option<&str>,
+        launch_cgroups: bool,
+        gpu_variable: Option<String>,
+    ) {
+        let replace_keys = [
+            ("%U", ""), // link(s)
+            ("%u", ""),
+            ("%F", ""), // files(s)
+            ("%f", ""),
+            ("%D", ""), // Deprecated
+            ("%d", ""),
+            ("%N", ""),
+            ("%n", ""),
+            ("%v", ""),
+            ("%m", ""),
+            ("%i", &self.icon.clone().unwrap_or_default()), // icon TODO: clone!
+            ("%c", &self.name),                             // name (translated)
+            ("%k", ""),                                     // filename as uri > file > none
+        ];
+        let mut command_string = self.exec.clone();
+        for replace_key in replace_keys {
+            command_string = command_string.replace(replace_key.0, replace_key.1)
+        }
+        let mut command: Vec<String> = Shlex::new(&command_string).collect();
+
+        if self.terminal {
+            if let Some(term) = term_command {
+                let command_string = term.to_string().replace("{}", &command_string);
+                command = Shlex::new(&command_string).collect();
+            } else if let Some(term) = std::env::var_os("TERMINAL") {
+                let term = term.into_string().expect("couldn't convert to string");
+                let mut command_new = vec![term, "-e".into()];
+                command_new.extend(command);
+                command = command_new;
+            } else {
+                return;
+            };
+        }
+        if launch_cgroups {
+            // TODO: clone
+            // info.id.clone().truncate(info.id.len() - ".desktop".len()); // remove .desktop extension
+            let parsed = Command::new("systemd-escape")
+                .arg(&self.id)
+                .output()
+                .unwrap()
+                .stdout;
+            let unit = format!(
+                "--unit=app-sirula-{}-{}",
+                String::from_utf8_lossy(&parsed).trim(),
+                id()
+            );
+            let mut command_new: Vec<String> = vec![
+                "systemd-run".into(),
+                "--scope".into(),
+                "--user".into(),
+                unit,
+            ];
+            command_new.extend(command);
+            command = command_new;
+        }
+
+        let mut exec = Command::new(&command[0]);
+        let mut exec = exec.args(&command[1..]);
+        if let Some(dir) = &self.path {
+            exec = exec.current_dir(dir)
+        }
+        if self.prefers_nondefault_gpu {
+            if let Some(prime) = gpu_variable {
+                exec = exec.env(prime, "1")
+            }
+        }
+
+        exec.spawn().expect("Error launching app");
     }
 }
